requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pandas==2.1.3
pyodbc==4.0.39
python-dotenv==1.0.0
google-generativeai==0.3.2
matplotlib==3.8.2
seaborn==0.13.1
streamlit==1.31.1 